+++ public/templates/show_album_group_disks.inc.php
@@ -90,9 +90,6 @@ $zipHandler = $dic->get(ZipHandlerInterface::class);
     <span id="rating_<?php echo $album->id; ?>_album">
         <?php echo Rating::show($album->id, 'album'); ?>
     </span>
-    <?php
-        } ?>
-    <?php if (AmpConfig::get('userflags')) { ?>
     <span id="userflag_<?php echo $album->id; ?>_album">
         <?php echo Userflag::show($album->id, 'album'); ?>
     </span>