+++ public/templates/show_videos.inc.php
@@ -32,7 +32,7 @@ use Ampache\Module\Util\Ui;
 /** @var array $object_ids */
 
 $web_path     = AmpConfig::get('web_path');
-$show_ratings = User::is_registered() && (AmpConfig::get('ratings') || AmpConfig::get('userflags'));
+$show_ratings = User::is_registered() && (AmpConfig::get('ratings'));
 $hide_genres  = AmpConfig::get('hide_genres');
 $is_table     = $browse->is_grid_view();
 //mashup and grid view need different css