+++ src/Module/Api/Method/FlagMethod.php
@@ -55,8 +55,8 @@ final class FlagMethod
      */
     public static function flag(array $input)
     {
-        if (!AmpConfig::get('userflags')) {
-            Api::error(T_('Enable: userflags'), '4703', self::ACTION, 'system', $input['api_format']);
+        if (!AmpConfig::get('ratings')) {
+            Api::error(T_('Enable: ratings'), '4703', self::ACTION, 'system', $input['api_format']);
 
             return false;
         }